+++ src/table_new/Ellipsis-new.tsx
